third_party/WebKit/LayoutTests/platform/win/http/tests/security/mixedContent/blocked-blob-url-script-in-data-iframe.https-expected.txt

CONSOLE ERROR: Not allowed to load local resource: blob:null/08d80dc4-044c-459d-94ed-704b794f0ffa
This is a testharness.js-based test.
PASS 'data:' URL frames cannot load 'blob:' resources originating from another data:-URL. 
Harness: the test ran to completion.
